Implements OAuth (using the third party signpost library to sign requests).
The server is now capable of obtaining and storing credentials in the appengine
datastore for a particular UserAccount for authorization to Buzz. I haven't
yet tried it with other OAuth based APIs, but this code should work fairly
generically.

Live at http://islay-test.appspot.com. Log in using your google account,
then click the "Connect to Buzz" button. You'll be redirected to an
authorization page, and if successful, back to our app with a success
message.