Eliminates the separate page for setting up network connections, and now just embeds it as a widget inside the settings link at the top right of the browser page.
This commit also changes the oauth flow a little so that when oauth calls us back,
we process the callback on the server side (via a custom servlet) rather than on
the client side, then redirect back to the client. This is more secure and also
fixes the redirection issue we were having with facebook (since we bypass the
need to set a gwt URL as the callback).

This commit also abstracts out the animation code so that it can be used elsewhere,
and combines the functionality of UserBar, PageHeader, NetworksPresenter into a new
single PresenterWidget, SettingsPresenter.

RecentPostsView is now effectively the "main page" of the application.